requirements.txt
requests
python-dotenv
openai>=1.0.0
pandas
numpy
